# improve performance of instance_norm